Separate jump predictions from sweep prediction provider
