+++ crates/zeta/src/sweep_ai.rs
@@ -1,10 +1,274 @@
-use std::fmt;
-use std::{path::Path, sync::Arc};
-
+use anyhow::{Context as _, Result};
+use cloud_llm_client::predict_edits_v3::Event;
+use futures::AsyncReadExt as _;
+use gpui::{
+    App, AppContext as _, Entity, Task,
+    http_client::{self, AsyncBody, Method},
+};
+use language::{Buffer, BufferSnapshot, Point, ToOffset as _, ToPoint as _};
+use lsp::DiagnosticSeverity;
+use project::{Project, ProjectPath};
 use serde::{Deserialize, Serialize};
+use std::{
+    collections::VecDeque,
+    fmt::{self, Write as _},
+    ops::Range,
+    path::Path,
+    sync::Arc,
+    time::Instant,
+};
+use util::ResultExt as _;
+
+use crate::{EditPrediction, EditPredictionId, EditPredictionInputs};
+
+const SWEEP_API_URL: &str = "https://autocomplete.sweep.dev/backend/next_edit_autocomplete";
+
+pub struct SweepAi {
+    pub api_token: Option<String>,
+    pub debug_info: Arc<str>,
+}
+
+impl SweepAi {
+    pub fn new(cx: &App) -> Self {
+        SweepAi {
+            api_token: std::env::var("SWEEP_AI_TOKEN")
+                .context("No SWEEP_AI_TOKEN environment variable set")
+                .log_err(),
+            debug_info: debug_info(cx),
+        }
+    }
+
+    pub fn request_prediction_with_sweep(
+        &self,
+        project: &Entity<Project>,
+        active_buffer: &Entity<Buffer>,
+        snapshot: BufferSnapshot,
+        position: language::Anchor,
+        events: Vec<Arc<Event>>,
+        recent_paths: &VecDeque<ProjectPath>,
+        diagnostic_search_range: Range<Point>,
+        cx: &mut App,
+    ) -> Task<Result<Option<EditPrediction>>> {
+        let debug_info = self.debug_info.clone();
+        let Some(api_token) = self.api_token.clone() else {
+            return Task::ready(Ok(None));
+        };
+        let full_path: Arc<Path> = snapshot
+            .file()
+            .map(|file| file.full_path(cx))
+            .unwrap_or_else(|| "untitled".into())
+            .into();
+
+        let project_file = project::File::from_dyn(snapshot.file());
+        let repo_name = project_file
+            .map(|file| file.worktree.read(cx).root_name_str())
+            .unwrap_or("untitled")
+            .into();
+        let offset = position.to_offset(&snapshot);
+
+        let recent_buffers = recent_paths.iter().cloned();
+        let http_client = cx.http_client();
+
+        let recent_buffer_snapshots = recent_buffers
+            .filter_map(|project_path| {
+                let buffer = project.read(cx).get_open_buffer(&project_path, cx)?;
+                if active_buffer == &buffer {
+                    None
+                } else {
+                    Some(buffer.read(cx).snapshot())
+                }
+            })
+            .take(3)
+            .collect::<Vec<_>>();
+
+        let cursor_point = position.to_point(&snapshot);
+        let buffer_snapshotted_at = Instant::now();
+
+        let result = cx.background_spawn({
+            let snapshot = snapshot.clone();
+            let diagnostic_search_range = diagnostic_search_range.clone();
+            async move {
+                let text = snapshot.text();
+
+                let mut recent_changes = String::new();
+                for event in &events {
+                    write_event(event.as_ref(), &mut recent_changes).unwrap();
+                }
+
+                let mut file_chunks = recent_buffer_snapshots
+                    .into_iter()
+                    .map(|snapshot| {
+                        let end_point = Point::new(30, 0).min(snapshot.max_point());
+                        FileChunk {
+                            content: snapshot.text_for_range(Point::zero()..end_point).collect(),
+                            file_path: snapshot
+                                .file()
+                                .map(|f| f.path().as_unix_str())
+                                .unwrap_or("untitled")
+                                .to_string(),
+                            start_line: 0,
+                            end_line: end_point.row as usize,
+                            timestamp: snapshot.file().and_then(|file| {
+                                Some(
+                                    file.disk_state()
+                                        .mtime()?
+                                        .to_seconds_and_nanos_for_persistence()?
+                                        .0,
+                                )
+                            }),
+                        }
+                    })
+                    .collect::<Vec<_>>();
+
+                let diagnostic_entries =
+                    snapshot.diagnostics_in_range(diagnostic_search_range, false);
+                let mut diagnostic_content = String::new();
+                let mut diagnostic_count = 0;
+
+                for entry in diagnostic_entries {
+                    let start_point: Point = entry.range.start;
+
+                    let severity = match entry.diagnostic.severity {
+                        DiagnosticSeverity::ERROR => "error",
+                        DiagnosticSeverity::WARNING => "warning",
+                        DiagnosticSeverity::INFORMATION => "info",
+                        DiagnosticSeverity::HINT => "hint",
+                        _ => continue,
+                    };
+
+                    diagnostic_count += 1;
+
+                    writeln!(
+                        &mut diagnostic_content,
+                        "{} at line {}: {}",
+                        severity,
+                        start_point.row + 1,
+                        entry.diagnostic.message
+                    )?;
+                }
+
+                if !diagnostic_content.is_empty() {
+                    file_chunks.push(FileChunk {
+                        file_path: format!("Diagnostics for {}", full_path.display()),
+                        start_line: 0,
+                        end_line: diagnostic_count,
+                        content: diagnostic_content,
+                        timestamp: None,
+                    });
+                }
+
+                let request_body = AutocompleteRequest {
+                    debug_info,
+                    repo_name,
+                    file_path: full_path.clone(),
+                    file_contents: text.clone(),
+                    original_file_contents: text,
+                    cursor_position: offset,
+                    recent_changes: recent_changes.clone(),
+                    changes_above_cursor: true,
+                    multiple_suggestions: false,
+                    branch: None,
+                    file_chunks,
+                    retrieval_chunks: vec![],
+                    recent_user_actions: vec![],
+                    // TODO
+                    privacy_mode_enabled: false,
+                };
+
+                let mut buf: Vec<u8> = Vec::new();
+                let writer = brotli::CompressorWriter::new(&mut buf, 4096, 11, 22);
+                serde_json::to_writer(writer, &request_body)?;
+                let body: AsyncBody = buf.into();
+
+                let inputs = EditPredictionInputs {
+                    events,
+                    included_files: vec![cloud_llm_client::predict_edits_v3::IncludedFile {
+                        path: full_path.clone(),
+                        max_row: cloud_llm_client::predict_edits_v3::Line(snapshot.max_point().row),
+                        excerpts: vec![cloud_llm_client::predict_edits_v3::Excerpt {
+                            start_line: cloud_llm_client::predict_edits_v3::Line(0),
+                            text: request_body.file_contents.into(),
+                        }],
+                    }],
+                    cursor_point: cloud_llm_client::predict_edits_v3::Point {
+                        column: cursor_point.column,
+                        line: cloud_llm_client::predict_edits_v3::Line(cursor_point.row),
+                    },
+                    cursor_path: full_path.clone(),
+                };
+
+                let request = http_client::Request::builder()
+                    .uri(SWEEP_API_URL)
+                    .header("Content-Type", "application/json")
+                    .header("Authorization", format!("Bearer {}", api_token))
+                    .header("Connection", "keep-alive")
+                    .header("Content-Encoding", "br")
+                    .method(Method::POST)
+                    .body(body)?;
+
+                let mut response = http_client.send(request).await?;
+
+                let mut body: Vec<u8> = Vec::new();
+                response.body_mut().read_to_end(&mut body).await?;
+
+                let response_received_at = Instant::now();
+                if !response.status().is_success() {
+                    anyhow::bail!(
+                        "Request failed with status: {:?}\nBody: {}",
+                        response.status(),
+                        String::from_utf8_lossy(&body),
+                    );
+                };
+
+                let response: AutocompleteResponse = serde_json::from_slice(&body)?;
+
+                let old_text = snapshot
+                    .text_for_range(response.start_index..response.end_index)
+                    .collect::<String>();
+                let edits = language::text_diff(&old_text, &response.completion)
+                    .into_iter()
+                    .map(|(range, text)| {
+                        (
+                            snapshot.anchor_after(response.start_index + range.start)
+                                ..snapshot.anchor_before(response.start_index + range.end),
+                            text,
+                        )
+                    })
+                    .collect::<Vec<_>>();
+
+                anyhow::Ok((
+                    response.autocomplete_id,
+                    edits,
+                    snapshot,
+                    response_received_at,
+                    inputs,
+                ))
+            }
+        });
+
+        let buffer = active_buffer.clone();
+
+        cx.spawn(async move |cx| {
+            let (id, edits, old_snapshot, response_received_at, inputs) = result.await?;
+            anyhow::Ok(
+                EditPrediction::new(
+                    EditPredictionId(id.into()),
+                    &buffer,
+                    &old_snapshot,
+                    edits.into(),
+                    buffer_snapshotted_at,
+                    response_received_at,
+                    inputs,
+                    cx,
+                )
+                .await,
+            )
+        })
+    }
+}
 
 #[derive(Debug, Clone, Serialize)]
-pub struct AutocompleteRequest {
+struct AutocompleteRequest {
     pub debug_info: Arc<str>,
     pub repo_name: String,
     pub branch: Option<String>,
@@ -22,7 +286,7 @@ pub struct AutocompleteRequest {
 }
 
 #[derive(Debug, Clone, Serialize)]
-pub struct FileChunk {
+struct FileChunk {
     pub file_path: String,
     pub start_line: usize,
     pub end_line: usize,
@@ -31,7 +295,7 @@ pub struct FileChunk {
 }
 
 #[derive(Debug, Clone, Serialize)]
-pub struct RetrievalChunk {
+struct RetrievalChunk {
     pub file_path: String,
     pub start_line: usize,
     pub end_line: usize,
@@ -40,7 +304,7 @@ pub struct RetrievalChunk {
 }
 
 #[derive(Debug, Clone, Serialize)]
-pub struct UserAction {
+struct UserAction {
     pub action_type: ActionType,
     pub line_number: usize,
     pub offset: usize,
@@ -51,7 +315,7 @@ pub struct UserAction {
 #[allow(dead_code)]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum ActionType {
+enum ActionType {
     CursorMovement,
     InsertChar,
     DeleteChar,
@@ -60,7 +324,7 @@ pub enum ActionType {
 }
 
 #[derive(Debug, Clone, Deserialize)]
-pub struct AutocompleteResponse {
+struct AutocompleteResponse {
     pub autocomplete_id: String,
     pub start_index: usize,
     pub end_index: usize,
@@ -80,7 +344,7 @@ pub struct AutocompleteResponse {
 
 #[allow(dead_code)]
 #[derive(Debug, Clone, Deserialize)]
-pub struct AdditionalCompletion {
+struct AdditionalCompletion {
     pub start_index: usize,
     pub end_index: usize,
     pub completion: String,
@@ -90,7 +354,7 @@ pub struct AdditionalCompletion {
     pub finish_reason: Option<String>,
 }
 
-pub(crate) fn write_event(
+fn write_event(
     event: &cloud_llm_client::predict_edits_v3::Event,
     f: &mut impl fmt::Write,
 ) -> fmt::Result {
@@ -115,7 +379,7 @@ pub(crate) fn write_event(
     }
 }
 
-pub(crate) fn debug_info(cx: &gpui::App) -> Arc<str> {
+fn debug_info(cx: &gpui::App) -> Arc<str> {
     format!(
         "Zed v{version} ({sha}) - OS: {os} - Zed v{version}",
         version = release_channel::AppVersion::global(cx),
